backend/requirements.txt
# ParaFast - EMS Multi-Agent System
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
pydantic>=2.0.0
dicttoxml>=1.7.0
fpdf2>=2.7.0
resend>=0.8.0
langchain-openai>=0.0.5
langchain-core>=0.2.0
langgraph>=0.0.20
openai>=1.0.0
playwright>=1.40.0

# Voice Layer
httpx>=0.27.0
websockets>=12.0
